flight_mode_manager: move WeatherVane out
I found it very confusing to find the weather vane instance created
in the FlightModeManager instead of in the respective flight tasks which
actually make use of it. The enable/disable logic was spread all over
the place making it hard to understand where weather vane is actually
used and where not.

By moving it into the respective flight tasks (in this case
ManualPosition and Auto) the logic to enable/disable is in one place,
plus we don't need to deal with instantiating it dynamically but it's
just part of a task.

The only functional change that comes with this commit is - at least
from what I can tell - that weather vane is now usable also from
multicopters, and not just VTOLs. This comes from the fact that the
system_status was not easily available inside the flight tasks. However,
I think that anyway, it makes sense to have weather vane available for
non VTOLs.