commander: remove COM_POSCTL_NAVL
The param COM_POSCTL_NAVL allowed to configure the GPS failsafe for
POSCTL to be LAND instead of ALTCTL.

I think this option was added but likely never used. Removing it makes
it easier to move forward to a state where POSCTL is the default and it
will briefly switch back to ALTCTL if the position estimate is not good
enough. I'm sure that behaviour is most intuitive and what users are
expecting.