flight_mode_manager: no autogen for commands
This refactor is required because the autogeneration to handle command
only really works for commands that have exactly the right name, as
DO_ORBIT does, however, it already falls apart for the NAV_TAKEOFF
command.

Also, in general, I'm not very fond of auto-generated stuff when you can
do it manually stupid and simple.