flight_mode_manager: way to move on to next task
This adds a facility for a flight task to signal that is has finished
and we can move on to another task. However, we can't quite make this
work yet for takeoff because we don't know whether to move on to MISSION
or LOITER afterwards. I suspect we first need to refactor the auto
flight tasks a bit and split them up into LOITER and mission.