commander: sync arming/disarming from sysid 2 to 3
In order to arm and disarm in sync, we now send the arming command from
vehicle with sysid 2 to sysid 3. We use the force setting to make sure
arming doesn't fail because of any failing preflight tests and disarming
doesn't fail because of the land detector not having triggered landed
yet.

Signed-off-by: Julian Oes <julian@oes.ch>