Don't switch back to previous item
This fixes and issue where the mission logic switches back to the
previous item if the wind pushes us back far enough.

This then leads to the altitude switching back and forth with the
previous item and makes everything worse, when we just want to make sure
we go up or down.